content_inbox/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
feedparser>=6.0,<7.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
eval_type_backport>=0.2,<1.0
